refactor: restructure pipeline into entities/utils, ie stage, and clearer triplet naming
- Move entity utilities (lut, subtree_filter, link_xrefs) into entities/utils/
- Extract IE from triplets/ into pipeline/ie/ as a standalone pipeline stage
- Move knowledge_graph.py to pipeline root (central hub, not triplet-specific)
- Rename ontology files to entity-pair convention (food_food, chemical_chemical, disease_disease)
- Move flavor.py to postprocessing/ (entity enrichment, not triplet building)
- Remove dead legacy _metadata_new.json fallback code
- Update ruff per-file-ignores for renamed ie/constants.py
