+++ backend/kgc/src/pipeline/entities/utils/subtree_filter.py
