+++ backend/kgc/src/pipeline/ie/constants.py
