+++ backend/kgc/src/pipeline/ie/resolver.py
