+++ backend/kgc/tests/conftest.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pandas as pd
 import pytest
 from src.models.settings import KGCSettings
-from src.pipeline.triplets.knowledge_graph import KnowledgeGraph
+from src.pipeline.knowledge_graph import KnowledgeGraph
 from src.stores.entity_store import EntityStore
 from src.stores.schema import (
     FILE_ENTITIES,
