+++ backend/kgc/tests/test_subtree_filter.py
@@ -2,7 +2,10 @@
 
 import pandas as pd
 from src.config.corrections import OntologyRoots
-from src.pipeline.entities.subtree_filter import _compute_descendants, filter_sources
+from src.pipeline.entities.utils.subtree_filter import (
+    _compute_descendants,
+    filter_sources,
+)
 
 
 def _make_edges(pairs: list[tuple[str, str]], edge_type: str = "is_a") -> pd.DataFrame:
