+++ pyproject.toml
@@ -34,7 +34,7 @@ ignore = [
 "**/preprocessing/constants/greek_letters.py" = ["RUF001"]
 "**/preprocessing/constants/punctuations.py" = ["RUF001"]
 "**/tests/test_preprocessing_constants.py" = ["RUF001"]
-"**/triplets/ie_constants.py" = ["RUF001"]
+"**/ie/constants.py" = ["RUF001"]
 "**/tests/test_ie_loader.py" = ["RUF001"]
 
 [tool.ruff.format]
